requirements.txt
# Core dependencies
gradio>=4.0.0
torch>=2.0.0
numpy>=1.24.0
requests>=2.31.0
websockets>=12.0
asyncio>=3.4.3
python-dotenv>=1.0.0
pyjwt>=2.8.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1
pyaudio>=0.2.13

# WebRTC dependencies
aiortc>=1.5.0
aiohttp>=3.9.0
python-socketio>=5.10.0

# Model dependencies
transformers>=4.35.0
sentencepiece>=0.1.99
accelerate>=0.24.0

# Logging and monitoring
python-json-logger>=2.0.7
prometheus-client>=0.17.1 